Add files via upload
I've managed to fix the hamburger design however still not fully finished. I had also fixed an image with the issues which I wanted to implement but it went wrong so I could rid of the feature. I still have to finish off the report section and contact section but then after that my website will be complete.